Abstracting away access to File and RandomAccessFile
